make a refactoring
