+++ cmd/oras/manifest/delete.go
@@ -16,6 +16,7 @@ package manifest
 import (
 	"fmt"
 
+	"github.com/spf13/cobra"
 	"oras.land/oras/cmd/oras/internal/option"
 	"oras.land/oras/internal/errors"
 )
@@ -25,6 +26,37 @@ type deleteOptions struct {
 	option.Remote
 
 	targetRef string
+}
+
+func deleteCmd() *cobra.Command {
+	var opts deleteOptions
+	cmd := &cobra.Command{
+		Use:   "delete name[:tag|@digest]",
+		Short: "[Preview] Delete a manifest from remote registry",
+		Long: `[Preview] Delete a manifest from remote registry
+** This command is in preview and under development. **
+
+Example - Delete a manifest tagged with 'latest' from repository 'locahost:5000/hello':
+  oras manifest delete localhost:5000/hello:latest
+
+Example - Delete a manifest with digest '99e4703fbf30916f549cd6bfa9cdbab614b5392fbe64fdee971359a77073cdf9' from repository 'locahost:5000/hello':
+  oras manifest delete localhost:5000/hello@sha:99e4703fbf30916f549cd6bfa9cdbab614b5392fbe64fdee971359a77073cdf9
+
+Example - Delete a manifest from the insecure registry:
+  oras manifest delete localhost:5000/hello:latest
+`,
+		Args: cobra.ExactArgs(1),
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			return opts.ReadPassword()
+		},
+		RunE: func(_ *cobra.Command, args []string) error {
+			opts.targetRef = args[0]
+			return deleteManifest(opts)
+		},
+	}
+
+	option.ApplyFlags(&opts, cmd.Flags())
+	return cmd
 }
 
 func deleteManifest(opts deleteOptions) error {
